source/portal/requirements.txt
fastapi
uvicorn[standard]
httpx
bcrypt
python-jose[cryptography]
python-multipart
